Start restructuring the website architecture
The old website was usign a dynamic architecture where some javascript code
was used to fetch the page content from a backend.
This is needlessy complex and time-consuming both from a server and user
perspective.
What is more it does not work when somebody is surfing the web with javascript
disabled.

From now on the code will move to a static website generator which is ideal for
my use case.