Add build timestamp at the end of each page.
This is to solve a bug (or is it a feature?) of github pages: the
website is not reloaded, unless index.html is changed. With this commit
each webpage has a timestamp which is updated every time the website is
compiled, forcing github to reload the content.

CAVEAT: the commit history in the website will not reflect content
changes. I will try to fix this in some way.